+++ src/CoreEx/WebApis/WebApiPublisher.cs
@@ -53,7 +53,33 @@ namespace CoreEx.WebApis
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The corresponding <see cref="ExtendedStatusCodeResult"/> <see cref="IActionResult"/> where successful.</returns>
         /// <remarks>The <paramref name="request"/> must have an <see cref="HttpRequest.Method"/> or <see cref="HttpMethods.Post"/>.</remarks>
-        public async Task<IActionResult> PublishAsync<TValue>(HttpRequest request, string? eventName, Func<WebApiParam<TValue>, Task>? beforeEvent = null, Action<EventData, TValue>? eventModifier = null,
+        public Task<IActionResult> PublishAsync<TValue>(HttpRequest request, string? eventName, Func<WebApiParam<TValue>, CancellationToken, Task>? beforeEvent = null, Action<EventData, TValue>? eventModifier = null,
+            HttpStatusCode statusCode = HttpStatusCode.Accepted, OperationType operationType = OperationType.Unspecified, IValidator<TValue>? validator = null, CancellationToken cancellationToken = default)
+            => PublishInternalAsync(request, false, default!, eventName, beforeEvent, eventModifier, statusCode, operationType, validator, cancellationToken);
+
+        /// <summary>
+        /// Performs a <see cref="HttpMethods.Post"/> operation with a request JSON content value of <see cref="Type"/> <typeparamref name="TValue"/> that is to be published using the <see cref="EventPublisher"/>.
+        /// </summary>
+        /// <typeparam name="TValue">The request JSON content value <see cref="Type"/>.</typeparam>
+        /// <param name="request">The <see cref="HttpRequest"/>.</param>
+        /// <param name="value">The The value (already deserialized).</param>
+        /// <param name="eventName">The optional event destintion name (e.g. Queue or Topic name).</param>
+        /// <param name="beforeEvent">A function that enables the value to be processed/validated before the underlying event publishing logic is enacted.</param>
+        /// <param name="eventModifier">An action to enable the <see cref="EventData"/> instance to be modified prior to publish.</param>
+        /// <param name="statusCode">The <see cref="HttpStatusCode"/> where successful.</param>
+        /// <param name="operationType">The <see cref="OperationType"/>.</param>
+        /// <param name="validator">The <see cref="IValidator{T}"/> to validate the deserialized value.</param>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>The corresponding <see cref="ExtendedStatusCodeResult"/> <see cref="IActionResult"/> where successful.</returns>
+        /// <remarks>The <paramref name="request"/> must have an <see cref="HttpRequest.Method"/> or <see cref="HttpMethods.Post"/>.</remarks>
+        public Task<IActionResult> PublishAsync<TValue>(HttpRequest request, TValue value, string? eventName, Func<WebApiParam<TValue>, CancellationToken, Task>? beforeEvent = null, Action<EventData, TValue>? eventModifier = null,
+            HttpStatusCode statusCode = HttpStatusCode.Accepted, OperationType operationType = OperationType.Unspecified, IValidator<TValue>? validator = null, CancellationToken cancellationToken = default)
+            => PublishInternalAsync(request, true, value, eventName, beforeEvent, eventModifier, statusCode, operationType, validator, cancellationToken);
+
+        /// <summary>
+        /// Performs a <see cref="HttpMethods.Post"/> operation with a request JSON content value of <see cref="Type"/> <typeparamref name="TValue"/> that is to be published using the <see cref="EventPublisher"/>.
+        /// </summary>
+        private async Task<IActionResult> PublishInternalAsync<TValue>(HttpRequest request, bool useValue, TValue value, string? eventName, Func<WebApiParam<TValue>, CancellationToken, Task>? beforeEvent = null, Action<EventData, TValue>? eventModifier = null,
             HttpStatusCode statusCode = HttpStatusCode.Accepted, OperationType operationType = OperationType.Unspecified, IValidator<TValue>? validator = null, CancellationToken cancellationToken = default)
         {
             if (request == null)
@@ -64,15 +90,15 @@ namespace CoreEx.WebApis
 
             return await RunAsync(request, async (wap, ct) =>
             {
-                var vr = await request.ReadAsJsonValueAsync(JsonSerializer, true, validator, cancellationToken).ConfigureAwait(false);
-                if (vr.IsInvalid)
-                    return vr.ToBadRequestResult();
+                var (wapv, vex) = await ValidateValueAsync(wap, useValue, value, true, validator, cancellationToken).ConfigureAwait(false);
+                if (vex != null)
+                    return vex.ToResult();
 
                 if (beforeEvent != null)
-                    await beforeEvent(new WebApiParam<TValue>(wap, vr.Value)).ConfigureAwait(false);
+                    await beforeEvent(wapv!, ct).ConfigureAwait(false);
 
-                var @event = new EventData { Value = vr.Value };
-                eventModifier?.Invoke(@event, vr.Value);
+                var @event = new EventData { Value = wapv!.Value };
+                eventModifier?.Invoke(@event, wapv.Value!);
 
                 if (eventName == null)
                     EventPublisher.Publish(@event);
@@ -101,7 +127,35 @@ namespace CoreEx.WebApis
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The corresponding <see cref="ExtendedStatusCodeResult"/> <see cref="IActionResult"/> where successful.</returns>
         /// <remarks>The <paramref name="request"/> must have an <see cref="HttpRequest.Method"/> or <see cref="HttpMethods.Post"/>.</remarks>
-        public async Task<IActionResult> PublishAsync<TColl, TItem>(HttpRequest request, string? eventName, Func<WebApiParam<TColl>, Task>? beforeEvents = null, Action<EventData, TItem>? eventModifier = null, int? maxCollSize = null,
+        public Task<IActionResult> PublishAsync<TColl, TItem>(HttpRequest request, string? eventName, Func<WebApiParam<TColl>, CancellationToken, Task>? beforeEvents = null, Action<EventData, TItem>? eventModifier = null, int? maxCollSize = null,
+            HttpStatusCode statusCode = HttpStatusCode.Accepted, OperationType operationType = OperationType.Unspecified, IValidator<TColl>? validator = null, CancellationToken cancellationToken = default) where TColl : IEnumerable<TItem>
+            => PublishInternalAsync(request, false, default!, eventName, beforeEvents, eventModifier, maxCollSize, statusCode, operationType, validator, cancellationToken);
+
+        /// <summary>
+        /// Performs a <see cref="HttpMethods.Post"/> operation with a request JSON content value of <see cref="Type"/> <typeparamref name="TColl"/> where each item is to be published using the <see cref="EventPublisher"/>.
+        /// </summary>
+        /// <typeparam name="TColl">The collection <see cref="Type"/></typeparam>
+        /// <typeparam name="TItem">The collection item <see cref="Type"/>.</typeparam>
+        /// <param name="request">The <see cref="HttpRequest"/>.</param>
+        /// <param name="value">The The value (already deserialized).</param>
+        /// <param name="eventName">The optional event destintion name (e.g. Queue or Topic name).</param>
+        /// <param name="beforeEvents">A function that enables the collection to be processed/validated before the underlying event publishing logic is enacted.</param>
+        /// <param name="eventModifier">An action to enable each item <see cref="EventData"/> instance to be modified prior to publish.</param>
+        /// <param name="maxCollSize">Overrides the default (see <see cref="SettingsBase.MaxPublishCollSize"/>) maximum publish collection size.</param>
+        /// <param name="statusCode">The <see cref="HttpStatusCode"/> where successful.</param>
+        /// <param name="operationType">The <see cref="OperationType"/>.</param>
+        /// <param name="validator">The <see cref="IValidator{T}"/> to validate the deserialized value.</param>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>The corresponding <see cref="ExtendedStatusCodeResult"/> <see cref="IActionResult"/> where successful.</returns>
+        /// <remarks>The <paramref name="request"/> must have an <see cref="HttpRequest.Method"/> or <see cref="HttpMethods.Post"/>.</remarks>
+        public Task<IActionResult> PublishAsync<TColl, TItem>(HttpRequest request, TColl value, string? eventName, Func<WebApiParam<TColl>, CancellationToken, Task>? beforeEvents = null, Action<EventData, TItem>? eventModifier = null, int? maxCollSize = null,
+            HttpStatusCode statusCode = HttpStatusCode.Accepted, OperationType operationType = OperationType.Unspecified, IValidator<TColl>? validator = null, CancellationToken cancellationToken = default) where TColl : IEnumerable<TItem>
+            => PublishInternalAsync(request, true, value, eventName, beforeEvents, eventModifier, maxCollSize, statusCode, operationType, validator, cancellationToken);
+
+        /// <summary>
+        /// Performs a <see cref="HttpMethods.Post"/> operation with a request JSON content value of <see cref="Type"/> <typeparamref name="TColl"/> where each item is to be published using the <see cref="EventPublisher"/>.
+        /// </summary>
+        private async Task<IActionResult> PublishInternalAsync<TColl, TItem>(HttpRequest request, bool useValue, TColl value, string? eventName, Func<WebApiParam<TColl>, CancellationToken, Task>? beforeEvents = null, Action<EventData, TItem>? eventModifier = null, int? maxCollSize = null,
             HttpStatusCode statusCode = HttpStatusCode.Accepted, OperationType operationType = OperationType.Unspecified, IValidator<TColl>? validator = null, CancellationToken cancellationToken = default) where TColl : IEnumerable<TItem>
         {
             if (request == null)
@@ -112,15 +166,15 @@ namespace CoreEx.WebApis
 
             return await RunAsync(request, async (wap, ct) =>
             {
-                var vr = await request.ReadAsJsonValueAsync(JsonSerializer, true, validator, ct).ConfigureAwait(false);
-                if (vr.IsInvalid)
-                    return vr.ToBadRequestResult();
+                var (wapv, vex) = await ValidateValueAsync(wap, useValue, value, true, validator, cancellationToken).ConfigureAwait(false);
+                if (vex != null)
+                    return vex.ToResult();
 
                 if (beforeEvents != null)
-                    await beforeEvents(new WebApiParam<TColl>(wap, vr.Value)).ConfigureAwait(false);
+                    await beforeEvents(wapv!, ct).ConfigureAwait(false);
 
                 var max = maxCollSize ?? Settings.MaxPublishCollSize;
-                var count = vr.Value.Count();
+                var count = wapv!.Value.Count();
                 if (count > max)
                 {
                     Logger.LogWarning("The publish collection contains {EventsCount} items where only a maximum size of {MaxCollSize} is supported; request has been rejected.", count, max);
@@ -130,7 +184,7 @@ namespace CoreEx.WebApis
                 if (count == 0)
                     return new AcceptedResult();
 
-                foreach (var item in vr.Value)
+                foreach (var item in wapv.Value!)
                 {
                     var ed = new EventData { Value = item };
                     eventModifier?.Invoke(ed, item);